prepare the data of the Mocker

+++ app/src/main/java/com/example/tarekbaz/watch_up/Models/Person.kt
@@ -2,10 +2,10 @@ package com.example.tarekbaz.watch_up.Models
 
 data class Person(val name: String,
                   val picture: Int,
-                  val birthDay : String,
-                  val bibliography : String,
-                  val movies : List<Movie>,
-                  val evaluation : Float,
-                  val comments : List<Comment>
+                  val birthDay: String,
+                  val bibliography: String,
+                  val movies: List<Movie>,
+                  val evaluation: Double,
+                  val comments: List<Comment>
                   ) {
 }
